Fix unnecessarily required fields
